Add a new script that mints some SDL to ChildGauge
